pyproject.toml
[project]
name = "codesearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "argparse>=1.4.0",
    "chromadb>=1.1.0",
    "einops>=0.8.1",
    "flash-attention>=1.0.0",
    "flask>=3.1.2",
    "gitpython>=3.1.45",
    "onnx>=1.19.0",
    "onnxruntime>=1.23.0",
    "optimum[onnxruntime]>=1.27.0",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]
